Several changes.  Some naming fixing.  The sky has been completed recoded and it works pretty well now.  There is a timing and sols that are counted as cycles of 10000.  There is a huge refactoring ahead.

+++ src/usercontrols.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <iostream>
 
-#define CONTROLLING_NONE -1
+#define CONTROLLING_NONE (size_t)-1
 
 struct controlregister
 {
